+++ src/main.py
@@ -1,56 +1,11 @@
-from flask import Flask, jsonify, request
-from datetime import datetime
 from pathlib import Path
 
-import sys, json
+import sys
 path_root = Path(__file__).parents[1]
-print(path_root)
 sys.path.append(str(path_root))
 
-from db.Store import DB
-
-api = Flask(__name__)
-dbName = "test"
-
-dbContext = DB(dbName)
-
-# Routes
-@api.route('/worklog/all', methods=['GET'])
-def getAllLog():
-   return jsonify(dbContext.getAll())
-
-@api.route('/worklog/remove/<int:id>', methods=['DELETE'])
-def removeLog(id):
-   popedItem = dbContext.remove(id)
-   dbContext.save()
-   return popedItem
-
-@api.route('/worklog/<int:id>', methods=['GET'])
-def getById(id):
-   return jsonify(dbContext.getById(id))
-
-@api.route('/worklog/customer/<customer>', methods=['GET'])
-def getByCustomer(customer):
-   return jsonify(dbContext.getByObject({
-      "müşteri": customer
-   }))
-
-@api.route('/worklog/work/<work>', methods=['GET'])
-def getByWork(work):
-   return jsonify(dbContext.getByObject({
-      "work": work
-   }))
-
-@api.route('/worklog/<int:id>', methods=['PUT'])
-def update(id):
-   return jsonify(dbContext.update(id, json.loads(request.data)))
-
-@api.route('/worklog/new', methods=['POST'])
-def add():
-   dbContext.add(datetime.now(), json.loads(request.data))
-   dbContext.save()
-   return request.data
+from api.API import API
 
 if __name__ == '__main__':
-   api.debug = True
-   api.run() #go to http://localhost:5000/ to view the page.
+   api = API()
+   api.app.run() #go to http://localhost:5000/ to view the page.
